(PC-19327) refactor(favorites): extract favorites fixtures
