+++ src/features/favorites/api/useAddFavorite.test.tsx
@@ -5,10 +5,8 @@ import { View } from 'react-native'
 import { FavoriteResponse, OfferResponse } from 'api/gen'
 import { useAuthContext } from 'features/auth/context/AuthContext'
 import { FavoritesWrapper } from 'features/favorites/context/FavoritesWrapper'
-import {
-  addFavoriteJsonResponseSnap,
-  paginatedFavoritesResponseSnap,
-} from 'features/favorites/fixtures/favoritesResponse'
+import { paginatedFavoritesResponseSnap } from 'features/favorites/fixtures/paginatedFavoritesResponseSnap'
+import { favoriteResponseSnap } from 'features/favorites/fixtures/favoriteResponseSnap'
 import { offerResponseSnap } from 'features/offer/fixtures/offerResponse'
 import { env } from 'libs/environment'
 import { EmptyResponse } from 'libs/fetch'
@@ -80,7 +78,7 @@ function simulateBackend(options: Options = defaultOptions) {
       } else if (hasAddFavoriteError) {
         return res(ctx.status(422), ctx.json({}))
       } else {
-        return res(ctx.status(200), ctx.json(addFavoriteJsonResponseSnap))
+        return res(ctx.status(200), ctx.json(favoriteResponseSnap))
       }
     }),
     rest.delete<EmptyResponse>(
@@ -114,10 +112,10 @@ describe('useAddFavorite hook', () => {
 
     await waitFor(() => {
       expect(onSuccess).toBeCalledWith({
-        ...addFavoriteJsonResponseSnap,
+        ...favoriteResponseSnap,
         offer: {
-          ...addFavoriteJsonResponseSnap.offer,
-          date: addFavoriteJsonResponseSnap.offer.date,
+          ...favoriteResponseSnap.offer,
+          date: favoriteResponseSnap.offer.date,
         },
       })
     })
