+++ src/features/favorites/components/Favorite.native.test.tsx
@@ -4,8 +4,9 @@ import { Share } from 'react-native'
 
 import { navigate } from '__mocks__/@react-navigation/native'
 import { api } from 'api/api'
-import { ExpenseDomain, FavoriteResponse, SubcategoryIdEnum, UserProfileResponse } from 'api/gen'
+import { ExpenseDomain, FavoriteResponse, UserProfileResponse } from 'api/gen'
 import { initialFavoritesState } from 'features/favorites/context/reducer'
+import { favoriteResponseSnap as favorite } from 'features/favorites/fixtures/favoriteResponseSnap'
 import { Credit } from 'features/user/helpers/useAvailableCredit'
 import { env } from 'libs/environment'
 import { EmptyResponse } from 'libs/fetch'
@@ -27,28 +28,7 @@ jest.mock('ui/components/snackBar/SnackBarContext', () => ({
 }))
 
 const credit: Credit = { amount: 100, isExpired: false }
-const favorite: FavoriteResponse = {
-  id: 393,
-  offer: {
-    coordinates: { latitude: 48.9263, longitude: 2.49008 },
-    date: null,
-    expenseDomains: [ExpenseDomain.all],
-    externalTicketOfficeUrl: 'https://externalbooking.test.com',
-    id: 146105,
-    subcategoryId: SubcategoryIdEnum.TELECHARGEMENT_MUSIQUE,
-    image: {
-      credit: null,
-      url: 'https://storage.gra.cloud.ovh.net/v1/AUTH_688df1e25bd84a48a3804e7fa8938085/storage-pc-dev/thumbs/mediations/CWMA',
-    },
-    name: 'Un lit sous une rivière',
-    price: null,
-    startDate: '2021-03-04T20:00:00',
-    startPrice: 270,
-    isSoldOut: false,
-    isExpired: false,
-    isReleased: true,
-  },
-}
+
 const user: UserProfileResponse = {
   isBeneficiary: true,
   bookedOffers: {},
