+++ src/features/favorites/components/FavoritesResults.native.test.tsx
@@ -6,7 +6,7 @@ import { FavoriteResponse, PaginatedFavoritesResponse } from 'api/gen'
 import { useFavorites, useRemoveFavorite } from 'features/favorites/api'
 import { FavoriteMutationContext } from 'features/favorites/api/types'
 import { initialFavoritesState } from 'features/favorites/context/reducer'
-import { paginatedFavoritesResponseSnap } from 'features/favorites/fixtures/favoritesResponse'
+import { paginatedFavoritesResponseSnap } from 'features/favorites/fixtures/paginatedFavoritesResponseSnap'
 import { env } from 'libs/environment'
 import { EmptyResponse } from 'libs/fetch'
 import { reactQueryProviderHOC } from 'tests/reactQueryProviderHOC'
