+++ src/features/favorites/fixtures/favoriteOfferResponseSnap.ts
@@ -1,0 +1,6 @@
+import { FavoriteOfferResponse } from 'api/gen'
+import { favoriteResponseSnap } from 'features/favorites/fixtures/favoriteResponseSnap'
+
+export const favoriteOfferResponseSnap: FavoriteOfferResponse = {
+  ...favoriteResponseSnap.offer,
+}
