+++ src/features/offer/components/AddToFavoritesButton/AddToFavoritesButton.native.test.tsx
@@ -3,10 +3,8 @@ import React from 'react'
 
 import { FavoriteResponse } from 'api/gen'
 import { useAuthContext } from 'features/auth/context/AuthContext'
-import {
-  paginatedFavoritesResponseSnap,
-  addFavoriteJsonResponseSnap,
-} from 'features/favorites/fixtures/favoritesResponse'
+import { paginatedFavoritesResponseSnap } from 'features/favorites/fixtures/paginatedFavoritesResponseSnap'
+import { favoriteResponseSnap } from 'features/favorites/fixtures/favoriteResponseSnap'
 import { AddToFavoritesButton } from 'features/offer/components/AddToFavoritesButton/AddToFavoritesButton'
 import { env } from 'libs/environment'
 import { EmptyResponse } from 'libs/fetch'
@@ -37,7 +35,7 @@ describe('<AddToFavoriteButton />', () => {
 
   it('should add favorite', async () => {
     const { getByText } = renderButton({
-      offerId: addFavoriteJsonResponseSnap.offer.id,
+      offerId: favoriteResponseSnap.offer.id,
     })
 
     fireEvent.press(getByText('Mettre en favori'))
@@ -76,7 +74,7 @@ const renderButton = (options?: Options) => {
         return res(ctx.status(415), ctx.json({}))
       } else {
         mockPostFavorite()
-        return res(ctx.status(200), ctx.json(addFavoriteJsonResponseSnap))
+        return res(ctx.status(200), ctx.json(favoriteResponseSnap))
       }
     })
   )
